pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "livins-report-agent"
version = "0.1.0"
description = "ReAct agent for property data analysis and report generation"
requires-python = ">=3.10"
dependencies = [
    # Core agent (same as rent_agent)
    "langgraph>=0.4",
    "langgraph-checkpoint>=2.0",
    "langchain-core>=0.3",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    # Report-specific
    "matplotlib>=3.9",
    "plotly>=6.0",
    "weasyprint>=62",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx",
    "ruff>=0.9",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/livins_report_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
